+++ mystery-deckbuilder/Assets/Scripts/Encounter/Effect.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: da090b209e801b641a4f711c52fb919e
+guid: e1a9b1bd42ad3fb4d88b8d4acb43466b
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
