+++ mystery-deckbuilder/Assets/Scripts/Encounter/Effect/Effect.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2c0b502a6bf6088439bc97567a9020db
+guid: f8d62236b2541b64a91345f5bd939702
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
