--- mystery-deckbuilder/Assets/Scripts/Encounter/Filters/Filter.cs
@@ -1,10 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public interface IFilter
-{
-    public string GetDescription();
-
-
-}
