--- mystery-deckbuilder/Assets/Scripts/Encounter/Filters/Filter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 023c78003e36f584db3e6e7e213860fe
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
